refactor(sdk): move gas-buffer helpers to utils/gas.ts (#1088)
Addresses review: keep the buffer logic out of client.ts.
