+++ sdk/src/gateway/utils/index.ts
@@ -1,4 +1,5 @@
 export * from './common';
 export * from './bitcoin-signer';
+export * from './gas';
 export * from './quote';
 export * from './tron';
